now mainline is the main notebook. added both twitter and wsj sentiment data in the model
