+++ r&d_ashok/README 1.md
@@ -1,0 +1,127 @@
+{
+ "cells": [
+  {
+   "cell_type": "markdown",
+   "id": "aa73b800-9274-4d4c-9b18-5b121d3dd11c",
+   "metadata": {},
+   "source": [
+    "# Project 2 - Cryvesto, Version 2\n",
+    "Utilize ML techniques to build a stock trading bot. \n",
+    "\n",
+    "---\n",
+    "\n",
+    "## User Story\n",
+    "A retail trader is so innundated with world news, business news, political events that he is unable to keep track of how these might impact trading. To help bring some sanity to this process, we are designing a bot which will utilize ML techniques and train the model with a sentiment analysis and how it impacted trading in the past. Use this knowledge to predict given the current news should an investor buy a stock or not.  \n",
+    "\n",
+    "The bot is capable of handling any stock symbol. If you give a particular stock symbol, it is going to utilize ML to give the best possible answer for that stock symbol\n",
+    "\n",
+    "---\n",
+    "\n",
+    "## Acceptance Criteria  \n",
+    "\n",
+    "The Cryvesto bot will use AWS Lex, Lambda and Sagemaker and must meet the following criteria:\n",
+    "\n",
+    "* TBD\n",
+    "\n",
+    "---\n",
+    "\n",
+    "## The Application  \n",
+    "\n",
+    "The application uses the following steps to build the robo advisor:\n",
+    "\n",
+    "### TBD\n",
+    "\n",
+    "\n",
+    "---\n",
+    "\n",
+    "## Technologies\n",
+    "The application is developed using:  \n",
+    "* Language: Python 3.7,   \n",
+    "* Packages/Libraries: Pandas; AWS Lex, AWS Lambda, AWS Sagemaker  \n",
+    "* Development Environment: VS Code and Terminal, Anaconda 2.1.1 with conda 4.11.0, Jupyterlab 3.2.9\n",
+    "* OS: Mac OS 12.1\n",
+    "\n",
+    "---\n",
+    "## Installation Guide\n",
+    "Following are the instructions to install the application from its Github respository.  \n",
+    "\n",
+    "### Clone the application code from Github as follows:\n",
+    "copy the URL link of the application from its Github repository      \n",
+    "open the Terminal window and clone as follows:  \n",
+    "\n",
+    "   1. %cd to_your_preferred_directory_where_you want_to_store_this_application  \n",
+    "    \n",
+    "   2. %git clone URL_link_that_was_copied_in_step_1_above   \n",
+    "    \n",
+    "   3. %ls     \n",
+    "        Project_2    \n",
+    "        \n",
+    "   4. %cd Project_2     \n",
+    "\n",
+    "At this point you will have the the entire application files in the current directory as follows (THIS IS JUST A SAMPLE):\n",
+    "\n",
+    "    * ageError                        (Test Result screenshot)  \n",
+    "    * bot w lambda                    (bot video)  \n",
+    "    * bot_without_lambda              (bot video)   \n",
+    "    * correctDialog                   (test result screenshot)  \n",
+    "    * incorrectAmt                    (test result screenshot)  \n",
+    "    * negativeAgeError                (test result screenshot)  \n",
+    "    * README.md                       (this file you are reading)  \n",
+    "    * Starter_Code  \n",
+    "        - Lambda  \n",
+    "            - lambda_function.py      (start code that was modified)\n",
+    "        - Test_Events                 (used for testing the Lambda function)\n",
+    "            - ageError.json  \n",
+    "            - correctDialog.json  \n",
+    "            - incorrectAmountError.json  \n",
+    "            - negativeAgeError.json  \n",
+    "       \n",
+    "---\n",
+    "\n",
+    "## Usage\n",
+    "The following details the instructions on how to run the application.  \n",
+    "\n",
+    "### Setup the environment and Run the application \n",
+    "\n",
+    "In order to run the Bot you will have to log on to the Amazon AWS Console and go through the process of creating the bot and then running the bot with lambda function as detailed above.\n",
+    "\n",
+    "---\n",
+    "\n",
+    "## Contributors\n",
+    "Ashok Pandey - ashok.pragati@gmail.com, www.linkedin.com/in/ashok-pandey-a7201237  \n",
+    "Dane Hayes - nydane1@gmail.com  \n",
+    "Scott Marler -scottjmarler@gmail.com  \n",
+    "Rensley Ramos -\n",
+    "Anna Joltaya\n",
+    "\n",
+    "---\n",
+    "\n",
+    "## License\n",
+    "The source code is the property of the developer. The users can copy and use the code freely but the developer is not responsible for any liability arising out of the code and its derivatives.\n",
+    "\n",
+    "---"
+   ]
+  }
+ ],
+ "metadata": {
+  "kernelspec": {
+   "display_name": "dev",
+   "language": "python",
+   "name": "dev"
+  },
+  "language_info": {
+   "codemirror_mode": {
+    "name": "ipython",
+    "version": 3
+   },
+   "file_extension": ".py",
+   "mimetype": "text/x-python",
+   "name": "python",
+   "nbconvert_exporter": "python",
+   "pygments_lexer": "ipython3",
+   "version": "3.7.11"
+  }
+ },
+ "nbformat": 4,
+ "nbformat_minor": 5
+}
